final out bonus

+++ ft_strchr.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strchr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lduthill <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: cgeoffra <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/07 16:50:58 by lduthill          #+#    #+#             */
-/*   Updated: 2023/02/15 15:49:35 by lduthill         ###   ########.fr       */
+/*   Created: 2023/02/02 15:06:32 by cgeoffra          #+#    #+#             */
+/*   Updated: 2023/02/02 15:28:42 by cgeoffra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 char	*ft_strchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
