+++ ft_striteri.c
@@ -3,17 +3,17 @@
 /*                                                        :::      ::::::::   */
 /*   ft_striteri.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lduthill <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: cgeoffra <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/02/10 14:55:43 by lduthill          #+#    #+#             */
-/*   Updated: 2023/02/15 18:32:39 by lduthill         ###   ########.fr       */
+/*   Created: 2023/02/15 10:59:22 by cgeoffra          #+#    #+#             */
+/*   Updated: 2023/02/15 11:00:41 by cgeoffra         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 void	ft_striteri(char *s, void (*f)(unsigned int, char*))
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
